fix:프로젝트 전반 오류수정

See merge request s06-final/S06P31C104!191

+++ frontend/src/pages/ProjectDetailPage/index.jsx
@@ -16,6 +16,7 @@ import {Button} from "@mui/material"
 import API from "../../api/API";
 import store from "../../utils/store";
 import markdownIt from "markdown-it";
+import swal from 'sweetalert';
 //pdf
 import { jsPDF } from "jspdf";
 import html2canvas from "html2canvas";
@@ -24,7 +25,9 @@ import html2canvas from "html2canvas";
 const DetailPage = () => {
   const id = useParams().projectId;
   let navigate = useNavigate();
-
+  
+  // 사용자 관리
+  const [myEmail, setMyEmail] = useState('');
   // 변수관리 hook
   const [category, setCategory] = useState('');
   const [title, setTitle] = useState('');
@@ -48,10 +51,24 @@ const DetailPage = () => {
   const [startWord, setStartword] = useState(-1)
   const [endWord, setEndword] = useState(-1)
   const [searchData, setSearchData] = useState([]);
+  const [isWriter, setIsWriter] = useState(false);
 
 
   useEffect(() => {
+    let myEmail = ''
+    let ptEmail = ''
+    // async function callmebaby() {
+    //   let token  = store.getToken()
+    //   const res = await API.get('/api/user', {header: token })
+    //   myEmail = res.data.user.email
+    //   console.log(res)
+    // }
     async function projectCall() {
+      const token  = store.getToken()
+      const response = await API.get('/api/user', { header: token })
+      myEmail = response.data.user.email
+      setMyEmail(response.data.user.email)
+
       const res = await API.get(`/api/project/${id}`)
       setCategory(res.data.project.category)
       setTitle(res.data.project.title)
@@ -66,10 +83,22 @@ const DetailPage = () => {
       setIntro(res.data.project.introduce)
       setThumbnail(res.data.project.thumbnail)
       setIsliked(res.data.project.isLiked)
-      setIsbookmarked(res.data.project.isBookmkared)
+      setIsbookmarked(res.data.project.isBookmarked)
+      console.log(res)
+      ptEmail = res.data.project.email
+      if (myEmail === ptEmail) {
+        setIsWriter(true)
+      }
     }
+    // callmebaby()
     projectCall()
+    // console.log(ptEmail)
+    // console.log(myEmail)
+    // if (ptEmail === myEmail) {
+    //   setIsWriter(true)
+    // }
   },[kai,id])
+  
   // let category = '자율'
   // let title = '라이키와 함께 자전거 여행을 떠나보세요!  나의 라이딩 메이트, RIKEY'
   // let stack = ['react-native','react','spring','엔진엑스','Unity','Unity']
@@ -115,6 +144,7 @@ const DetailPage = () => {
 
 
   const commentBox = comment.map((item) => {
+    console.log(item)
     const regex = /@.*[원|장]/
 
     let pingping = item.content
@@ -144,18 +174,39 @@ const DetailPage = () => {
         <div className="comment-content" id={item.commentId} dangerouslySetInnerHTML={{__html: pingping}}>
           
         </div>
-        <div>
-          <p className="project-detail-red" onClick={() => deleteComment(item.commentId)}>삭제하기</p>
-        </div>
+        { (myEmail === item.email) ?
+          <div>
+            <p className="project-detail-red" onClick={() => deleteComment(item.commentId)}>삭제하기</p>
+          </div>
+          :
+          null
+        }
       </div>
 
     </div>
   })
   const deleteProject = async() => {
-
-    store.getToken();
-    await API.delete(`/api/project/${id}`)
-    navigate('/')
+    swal({
+      title: "프로젝트 삭제",
+      text: "정말 프로젝트를 삭제하시겠습니까?",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          
+          swal("프로젝트가 성공적으로 삭제되었습니다.", {
+            icon: "success",
+          });
+          store.getToken();
+          API.delete(`/api/project/${id}`)
+          navigate('/')
+        } else {
+          swal("당신의 프로젝트가 살아남았습니다.");
+        }
+      });
+      
   }
 
 
@@ -236,7 +287,14 @@ const DetailPage = () => {
     }
   }
 
-
+  const receiveReadme = async() => {
+    try {
+      await API.put(`/api/project/${id}/readme`)
+      swal("갱신 완료","Read Me가 성공적으로 갱신되었습니다!","success")
+    }catch(e) {
+      console.log(e)
+    }
+  }
 
   async function searchWord(word) {
     const res  = await API.get(`/api/user-info/search?keyword=${word}`)
@@ -278,10 +336,46 @@ const DetailPage = () => {
     </div>
   });
   
+  const noneBepo = () => {
+    swal("데모 사이트", "배포서버가 준비되어 있지 않습니다.", "error")
+  }
+
+  const whichOne = () => {
+    swal("리뷰와 버그리포트중 무엇부터 확인하고 싶으세요?", {
+      buttons: {
+        review: {
+          text: "리뷰",
+          value: "review",
+        },
+        bugreport: {
+          text: "버그 리포트",
+          value: "bugreport",
+        },
+      },
+    })
+      .then((value) => {
+        switch (value) {
+          
+        case "bugreport":
+          navigate(`/project/${id}/opinions/report`)
+          break;
+     
+        case "review":
+          navigate(`/project/${id}/opinions/review`)
+          break;
+     
+        }
+      });
+  }
   return (
     
     <div className="project-div">
-      <img className="detailImage" src={detailImage} alt="detailImage" />
+      { (thumbnail) ?
+        <img className="thumbImage" src={thumbnail} alt="detailImage" />
+        :
+        <img className="detailImage" src={detailImage} alt="detailImage" />
+      }
+
       <div className="project-body-div">
         <div className="title-div">
           <div className="project-part">{category}</div>
@@ -292,31 +386,34 @@ const DetailPage = () => {
             <img className="icon" src={constructionPic} alt="conpic" />
             {stackBox}
           </span>
-
-          <span className="option-div">
-            <div className="option-category">ReadMe 갱신</div>
-            <div className="option-category">
-              <Link 
-                className="to-edit" 
-                to="edit"
-                state={{
-                  editTitle : title,
-                  editCategory: category,
-                  editStack: stack,
-                  editMember: member,
-                  editRepo : repo,
-                  editBepo : bepo,
-                  editReadme : readme,
-                  edittAuthormember : authormember,
-                  editIntro : intro,
-                  editThumbnail : thumbnail
-                }}
-              >
-                수정
-              </Link>
-            </div>
-            <div onClick={deleteProject} className="option-category-red">삭제</div>
-          </span>
+          { (isWriter) ?
+            <span className="option-div">
+              <div className="option-category" onClick={receiveReadme}>ReadMe 갱신</div>
+              <div className="option-category">
+                <Link 
+                  className="to-edit" 
+                  to="edit"
+                  state={{
+                    editTitle : title,
+                    editCategory: category,
+                    editStack: stack,
+                    editMember: member,
+                    editRepo : repo,
+                    editBepo : bepo,
+                    editReadme : readme,
+                    edittAuthormember : authormember,
+                    editIntro : intro,
+                    editThumbnail : thumbnail
+                  }}
+                >
+                  수정
+                </Link>
+              </div>
+              <div onClick={deleteProject} className="option-category-red">삭제</div>
+            </span>
+            :
+            null
+          }
         </div>
 
         <div className="member-div">
@@ -327,25 +424,36 @@ const DetailPage = () => {
         </div>
 
         <div className="git-div">
-          <a href="https://www.naver.com" rel="noreferrer" target='_blank' className="link-a">
+          <a href={repo} rel="noreferrer" target='_blank' className="link-a">
             <img className="icon" src={gitRepo} alt="gitRepo" />
             <span>
               Git Repo
             </span>
           </a>
-          <a href="https://www.kakao.com" rel="noreferrer" target='_blank' className="link-a">
-            <img className="icon" src={google} alt="google" />
-            <span>
-              Demo Site
+          { (bepo) ?
+            <a href={bepo} rel="noreferrer" target='_blank' className="link-a">
+              <img className="icon" src={google} alt="google" />
+              <span>
+                Demo Site
+              </span>
+            </a>
+            : 
+            <span onClick={noneBepo} rel="noreferrer" target='_blank' className="link-a">
+              <img className="icon" src={google} alt="google" />
+              <span>
+                Demo Site
+              </span>
             </span>
-          </a>
-          <Link 
+          }
+          {/* <Link 
             to={`/project/${id}/opinions/review`}
             className="link-a"
-          >
+          > */}
+          <span className="which-one" onClick={whichOne} >
             <img className="icon" src={report} alt="report" />
-            리뷰·버그 리포트
-          </Link>
+            <span>리뷰·버그 리포트</span>
+            {/* </Link> */}
+          </span>
         </div>
         <div>
           <button onClick={printDocument}>Print</button>
